+++ src/main/java/com/example/isproject/repository/PetRepo.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.reactive.ReactiveCrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PetRepo extends ReactiveCrudRepository<Pet,Integer>{
+public interface PetRepo extends ReactiveCrudRepository<Pet,Long>{
 }
